backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
mysql-connector-python==8.0.33
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.15.0
matplotlib==3.7.2
python-jose==3.3.0
email-validator==2.0.0.post2
flask-migrate==4.0.5
flask-sqlalchemy==3.0.3